feat: catcherror
