+++ frontend/src/app/Store.js
@@ -3,8 +3,8 @@ import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
 import { setupListeners } from "@reduxjs/toolkit/query";
 import userApi from "../pages/profile/UserApi";
-import authReducer from "../pages/loginSlice";
-import registerApi from "../pages/registerApi";
+import authReducer from "../pages/Login/loginSlice";
+import registerApi from "../pages/Register/registerApi";
 
 const persistConfig = {
     key: "root",
