+++ frontend/src/pages/Home/footer.jsx
