+++ frontend/src/pages/Passwords/PasswordReset.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import axios from 'axios';
 import './PasswordReset.scss';
-import BASE_URL from './baseUrl';
+import BASE_URL from '../baseUrl';
 
 const PasswordReset = () => {
   const { register, handleSubmit, formState: { errors, isSubmitting } } = useForm();
