+++ frontend/src/pages/Passwords/PasswordReset.scss
