+++ frontend/src/pages/dashboard/Sidebar.scss
