+++ frontend/src/pages/dashboard/dashboard.jsx
@@ -1,5 +1,5 @@
 import { useState, useCallback, createContext } from "react";
-import Sidebar from "./Sidebar.jsx";
+import Sidebar from "../dashboard/Sidebar";
 
 const SidebarContext = createContext();
