+++ frontend/src/pages/dashboard/dashboard.scss
