+++ frontend/src/pages/styles/Error.scss
